requirements.txt
streamlit
pillow
pdfplumber
google-generativeai
